Fix linter warnings

+++ backend/src/main/java/com/entregaFinal/gestion/service/ProductoService.java
@@ -19,14 +19,24 @@ public class ProductoService {
     }
 
     public Optional<Producto> getProductoById(String id) {
+        if (id == null) {
+            return Optional.empty();
+        }
         return productoRepository.findById(id);
     }
 
     public Producto addProducto(Producto producto) {
+        String id = producto.getId();
+        if (id != null && productoRepository.existsById(id)) {
+            return null; // No se puede agregar un producto con un ID que ya existe
+        }
         return productoRepository.save(producto);
     }
 
     public Producto updateProducto(String id, Producto producto) {
+        if (id == null) {
+            return null;
+        }
         if (productoRepository.existsById(id)) {
             producto.setId(id);
             return productoRepository.save(producto);
@@ -35,6 +45,10 @@ public class ProductoService {
     }
 
     public boolean deleteProducto(String id) {
+        if (id == null) {
+            return false;
+        }
+
         if (productoRepository.existsById(id)) {
             productoRepository.deleteById(id);
             return true;
